Spec: allow storing more than one dep from a single package
Introduce a new data structure, called _EdgeMap
that maps package names to list of DependencySpec
objects (edges in the DAG).

This data structure is used to track both
dependencies and dependents of a node. It
allows filtering the stored values by
parent or child name and by dependency types.

Unit tests added:
- Regression on 11983
- Synthetic construction of split dependency
- Synthetic construction of bootstrapping