containerize: pin the Spack version used in a container
This commit permits to specify the url and ref of
the Spack instance to be used in a container recipe.

If this is done an initial bootstrap stage is added,
which builds an image with Spack setup and ready to
install software.

A few new OSes have been added: centos:8, amazonlinux:2,
ubuntu:20.04, alpine:3.