seacas: Update recipe to find faodel in recent versions of seacas
This should fix #40237

Explciitly sets the CMake variables  Faodel_INCLUDE_DIRS and Faodel_LIBRARY_DIRS when +faodel.
This seems to be needed for recent versions of seacas (seacas@2021-04-02:), but should be safe
to do for all versions.

For Faodel_INCLUDE_DIRS, it looks like Faodel has header files under $(Faodel_Prefix)/include/faodel,
but seacas is not including the "faodel" part in #includes.  So add both $(Faodel_Prefix)/include
and $(Foadel_Prefix)/include/faodel

Also, it looks like a dependency on cereal was added in seacas@2021-04-02: +faodel.
I do not see it documented as an optional library, and looking at the install-tpl.sh script included
with seacas, there does not appear to be any flags to explicitly enable or disable the dependency,
but it is automatically added whenever faodel is selected.  Based on that, I am *not* adding a
+/-cereal variant, but adding cereal as a dependency for +faodel variants.