allow commands with unparsed arguments after --
Unlike unknown arguments, unparsed arguments are completely ignored by argument
parsing. The convention used is that any arguments after a -- are not
considered by the option parsing, but instead passed as additional parameter to
the command. This currently isn't considering the case where both unknown and
unparsed arguments could exist in a single command.