Config for Mac Sonoma 14.6.1 with Command Line Tools 15.3 (#471)
This PR adds in the necessary configuration to get the latest MacOS and Command Line Tools versions working (14.6.1 and 15.3 respectively).

This PR includes a patch for FMS 2024.02 and a variant spec that builds shared (dylib) libraries. This makes it unnecessary to do the manual FMS build workaround that was recently published. The FMS patch is only needed for version 2024.02 in that the next version (2024.03) contains the update to the CMakeLists.txt file that adds in the shared library option.

---------

Co-authored-by: Dom Heinzeller <dom.heinzeller@icloud.com>