Update ESMF from spack develop: add python dependency and extension (not for release/1.8.0) (#470)
Cherry-picking the changes from PR https://github.com/spack/spack/pull/45504/commits for our spack-stack-dev branch. With the exception of the "depends-on c/c++/fortran" changes (that our spack lib code doesn't support yet) and the disabling of OpenMP for macOS (only in our fork), the esmf package.py files are identical between spack develop and jcsda spack-stack-dev as of beginning of September 2024.