Fix errors found by running `spack audit externals`
I got the following errors when I ran the command with my cluster's
Spack installation:

```console
$ spack audit externals
PKG-EXTERNALS: 4 issues found
1. builtin.gcc: illegal regex in "gcc@9.4.0 languages='c,c++'" extra attributes
    .*/bin/g++ is not a valid regex
2. builtin.nvhpc: illegal regex in "nvhpc@20.9~blas~lapack~mpi" extra attributes
    .*/bin/nvc++ is not a valid regex
3. builtin.nvhpc: illegal regex in "nvhpc@20.9~blas~lapack~mpi" extra attributes
    .*/bin/nvc++ is not a valid regex
4. builtin.nvhpc: illegal regex in "nvhpc@20.9~blas~lapack~mpi" extra attributes
    .*/bin/nvc++ is not a valid regex
```

The problem was that `+` is part of the regex grammar, so it needs to be
escaped. I'm not sure why this isn't detected as part of global testing,
but it should still be fixed.

I'm not a fan of needing to write `".*/bin/g\\+\\+"` instead of
`'.*/bin/g\+\+'`, but neither of the YAML files I modified use
single-quoted strings at all, so I kept using double-quoted strings for
consistency.