Remove Spec.satisfies dependency on spack.repo.PATH
This change makes `Spec.satisfies` independent of the details of
the repositories being used. This also changes the semantic slightly,
as now propagating variants will be satisfied if not contradicted.

For instance:

  spec.satisfies("++mpi")

if it has no `~mpi` variant set in any node. It follows that a spec
with no `mpi` variants satisfies `++mpi`.