Spec: prefer a splice-specific method to `__len__`
Automatic splicing say `Spec` grow a `__len__` method but it's only used in one place
and it's not clear the semantics are useful elsewhere. It also runs the risk of Specs
one day being confused for other types of containers.

Rather than introduce a new function for one algorithm, let's use a more specific method
in the splice code.

- [x] introduce `node_count()` method in `_resolve_automatic_splices`
- [x] delete `Spec.__len__()` and `Spec.__bool__()`

Signed-off-by: Todd Gamblin <tgamblin@llnl.gov>