aws buildcache stack: Use specific path in `compiler add`
Use the specific argument to identify which compiler should be added. This does
not change the behavior in the stack (other than speeding it up due to shorter
search path) but helps when running this script on other OSs which might have
incomplete compiler installation we do not wish to add.

Also, explicitly conflict all packages with gcc@7 libraries. These libraries can
really wreck havoc on a build (using a more modern compiler) if they are added
in e.g. C++ link line.