Remove years from license headers
Spack is a Linux Foundation project now, and the guidance for copyrights
for LF projects can be found
[in this blog post](https://www.linuxfoundation.org/blog/blog/copyright-notices-in-open-source-software-projects).

In summary:

> many LF project communities recommend using a more general statement in a form similar to the following (where XYZ is the project’s name):
>
> Copyright The XYZ Authors.
> Copyright The XYZ Contributors.
> Copyright Contributors to the XYZ project.

I have updated Spack's license headers to mirror this:

```
```

We don't need to include years, so this will save us time updating thousands
of files every year. I didn't completely get rid of the old header -- I also
updated `COPYRIGHT` to include the fact that Spack was created in 2013 at LLNS,
LLC. So that information is still present--in one place.

- [x] Simplify license header and remove copyright years
- [x] Update all licensed files in Spack
- [x] Update `spack license verify`
- [x] Remove `spack license update-copyright-year` subcommand
- [x] Update `COPYRIGHT` file

Signed-off-by: Todd Gamblin <tgamblin@llnl.gov>