Assign priorities to config scopes
Up to now configuration behaved like a stack, where scopes could be
only be pushed at the top.

This commit allows to assign priorities to scopes, and ensures that
scopes of lower priorities are always "below" scopes of higher
priorities.

When scopes have the same priority, what matters is the insertion order.