Deprecate "make", etc. defined at module level
When we setup the build environment, we monkeypatch the package
module to have a "make", "gmake", and "ninja" attribute - regardless
of whether the package in question depends on them.

This commit deprecates those attributes, and makes any use of
them print a helpful error message, suggesting the user to add
a dependency on the corresponding package.