openmpi: do not pass --with-wrapper-ldflags
On macOS you cannot unconditionally pass `-rpath`, it can conflict with
`-r`.

I don't think mpicc is the place to inject rpaths to compiler runtime
libraries. That's up to the compiler's config files (spec files in gcc,
config files in llvm).

Also remove some patches that are redundant in newer versions of
openmpi.