relocate.py: don't warn about symlinks
`relocate_links` is warns when the target is absolute and not matched by
any prefix from the prefix to prefix map.

This can lead to false positives, cause the prefix to prefix map does
not contain trivial/identity elements whenever a package is installed to
its original location.

Since relocate_links is the odd one out there (we don't warns about
similar issues with rpaths, etc), just remove this since it's wrong.