env.dev_specs: remove caching for computed dev_specs
this cached value is not properly invalidated when develop config is
written, as in the `spack develop` command. Some accesses are accidentally
correct because a read or write transaction on the environment will
invalidate the `self._dev_specs` as a side-effect. Considered adding
invalidation but determined it was not worth the effort to bridge the gap in
interface between `spack.config` and `spack.environment`.