Add explicit path to zlib library to link to
CMake's find zlib looks for zlib before zdll. On Windows zlib is the static lib, and zdll the import library. LibPNG only links to shared zlib, but recieves zlib on Windows, resulting in a search for non import mangled symbols, which results in a linker failure.
The directly provides the correct library to the libpng CMake