python: allow it as a build-tool again
fixes #44289

Python was removed from being a build tool,
due to issues when reusing specs. This PR
adds a new rule to match the interpreter among
different Python packages, in clingo.

It also adds a bunch of new "build-tools", so that
specs like:

py-matplotlib backend=tkagg

can be concretized in one go.

Signed-off-by: Massimiliano Culpo <massimiliano.culpo@gmail.com>